cmake: automatic installation of preset files
Instead of statically add CMakeLists.txt files for all the presets,
change the root CMakeLists.txt of the presets directory to dynamically
scan for all the subdirectories and generate install rules for the
*.krp files.
This results in all the .krp files being installed automatically by
cmake, thus making the manual update with update-cmakelists-dot-txt.sh
no more needed (and that script has been removed).

The only small drawback is that when adding/removing a .krp file, for
cmake to notice that in an existing build directory it is needed to
trigger a build system update, which can be done just by touch'ing any
CMakeLists.txt file and rerunning the build.