fix stale pointer in AVProbeData structure
m_inputUrl.pathOrUrl() returns a new QString object, and calling
toLocal8Bit() on it will return a new QByteArray; but since none of them
is being stored in a variable, they will be destroyed, making the
const char* assigned to AVProbeData::filename point to deleted data.

Hence, store the QByteArray with the file name in a variable, so
AVProbeData::filename is valid during its existence.