lirc: fix possibly-invalid data passed to lirc_readconfig
Calling toUtf8() on a QString returns a new QByteArray; but since it is
not stored in a variable, it will be destroyed, making
lirc_cfg_filename_c point to deleted data (and thus the following strdup
may run into troubles).

Hence, store the QByteArray with the file name in a variable and stop
strdup'ing it, as apparently lirc_readconfig does not modify the "file"
parameter at all (why it hasn't been made const, then...).