alsa: use custom models for the devices comboboxes
Create custom models for the playback/capture devices/mixers, and use
them for the comboboxes in the ALSA configuration: these models cache
the results, and reset themselves only if the new list of
devices/mixers change (which will be only rarely).

Considering that the list of devices/mixers is reloaded every 10 secs,
this avoids continuously removing/adding items; this also workarounds
a leak in the qtaccessiblewidgets plugin of Qt4, which never deletes
the accessibility child objects when items are removed from item/views
(like the one used by QComboBox).