RadioStationListView: migrate to item/view
Create a new RadioStationModel which holds most of the logic for
handling a list of stations, including adding/removing items and
drag&drop (which does not require further logic anywhere else now,
making things much simplier).

Switch RadioStationListView to QTreeView + RadioStationModel,
simplifiying it a lot (with some APIs of it being just convenience
wrappers over RadioStationModel).

Adapt StationSelector the new RadioStationListView APIs, dropping the
manual code for drag&drop (automatically handled by the view), and
using a QModelIndex-based approach for moving rows from one list view
to the other.

The only notable change is that there is no more column with the item
numbers, as it is mostly an internal detail which can be confusing when
removing/adding stations. Except from that, everything else should work
as it used to be.