remove HAVE_CONFIG_H + <config.h> usage
Blocks like:
  #ifdef HAVE_CONFIG_H
  #include <config.h>
  #endif
comes from autotools times, where this is customary to do.
OTOH, currently there is no global configuration, and HAVE_CONFIG_H is
not even defined, so it is possible to get rid of these blocks all
around.

In case a config-foo.h is created later, will be better to include it
unconditionally and only where needed.