refactor the plugin loading
Since PluginBase is not QObject-based, we cannot use KPluginLoader and
KPluginFactory; make use of QPluginLoader by exposing a QObject-based
custom factory which exposes the information about the plugins using
KAboutData (so it possible to show them), replacing the custom exposed
symbols and their manual loading.

The only required changes for the users of PLUGIN_LIBRARY_FUNCTIONS and
PLUGIN_LIBRARY_FUNCTIONS2 is the usage of k1i8n instead of i18n, as it
is what KAboutData needs, and catalogs for the descriptions of plugins
may not be loaded yet.

The plugin system should work as before; the only difference is that
unloading a plugin at runtime won't actually unload it anymore, as it is
generally an unsafe operation.