cmake: centralize gettext .po handling
Create a process_po_files() macro to handle the rules to compile the .po
files in the source directory to .gmo and install them as .mo in the
correct place; based much on the existing (repeated in every po
subdirectory) code, but simplified. Make use of that macro in each
CMakeLists.txt.

Also, move the search for msgfmt in the root CMakeLists.txt, so there
is no need to repeat that elsewhere.