Reduce the number of plugin lists refresh
When building PluginManagerConfiguration, let only slotCancel() call
noticePluginLibrariesChanged(), which will call noticePluginsChanged()
later on, instead of calling them in other 2 places.

This reduces by 1 the times noticePluginLibrariesChanged() is called
(and thus that the list of plugin classes is cleared and re-filled),
and by 3 the times noticePluginsChanged() is called (and thus clearing
and re-filling the list of plugin instances).