Rework the plugin configuration page
The separation between plugin classes and instances in different tab is
slightly confusing, for example instances created in the classes tab
appears in a different tab... also, the libraries tab is the first,
while in reality is the least used (dealing with libraries should be
done less often than dealing with instances).

For this reason, rework a bit the plugin configuration page: unify the
classes and instances in a single tab, displaying them in a tree-like
structure (instances as children of each class). PluginsModel is the
model representing classes and instances, while PluginsDelegate is a
normal delegate which can render rich text.